requirements.txt
Pillow==9.2.0
timm==0.6.7
torch==1.12.1+cu116
torchvision==0.4.1
